httpd: fix compiler warning (cherry picked from commit 4e703a579d74587fe8f82fa040e3b39fe4b93b2d)

+++ modules/httpd/httpd_proc.c
@@ -212,7 +212,7 @@ static int post_iterator (void *cls,
 			" value=[%p]->'%s' off=%ld size=%ld\n",
 			cls, kind, key, key,
 			filename, content_type, transfer_encoding,
-			value, value, off, size);
+			value, value, (long long unsigned int)off, size);
 
 
 	pr = (struct post_request*)cls;
